Does a few things I was too lazy to split into multiple commits.
Adds a movement script to the StarFighter, giving it a Rigidbody with no gravity and moves forward at a constant speed.

Attempts to enable nullable reference types. This required upgrading to the beta release of Unity and setting C# to version 8 as well as setting the `Nullable` flag. This turned out to be far harder than I expected, as Unity generates the `.csproj` file where the flag should go. Unity doesn't seem to provide any means to set this flag and Visual Studio only shows warnings if I use `#nullable enable`.

I found that `csc.rsp` can set the flag and will show warnings in Unity without requiring `#nullable enable`. Unfortunately this still does not show the same warning in Visual Studio. It also does not make nullable issues hard errors, but only warnings. This can be configured in the `csc.rsp` file, but it causes errors for the entire compilation, rather than the specific project. Supposedly you can use `.asmdef` files to limit the flags to just this single assembly, but I wasn't able to get that to work.

For now, I'll just add `#nullable enable` to get Visual Studio to show the warnings, but it's really frustrating that I couldn't get this working the way I want it to.

Enables the new input system, haven't used it for anything yet.