Adds basic inputs.
This introduces simple turning controls, treating each direction as an axis and rotating the starfighter in that direction. It does not generate forward momentum since it has a constant velocity in the direction it is facing on start, but it does have basic steering in all three rotational axes.

For now, all the controls are via the keyboard, though pitch and yaw will likely be handled by the mouse eventually. Thrust and roll will mostly become the WASD keys, but this is a good starting point nonetheless.

Unfortunately the generated controls class has a nullability error I can't easily suppress. It's only one and not likely to multiply much so I'm ignoring it for now, but hopefully I can revisit this at some point in the future.