Updates StarFighter movement logic to use mouse for pitch and yaw control, moving roll control to A/D keys.
This brought several bugs to my attention:
1. I was setting the acceleration rather than adding a force, this had the effect of making the ship not respect angular momentum correctly.
2. The pitch/yaw/roll axes were world-relative, not StarFighter-relative, which cause some very strange movement.
3. The `performed` action for button presses is only on key down and does not repeatedly emit if the key is held. The StarFighter was not continually moving, it was moving once and then allowing momentum to take care of the rest. Updated these to use `started` and `canceled` actions which more accurately represent key down and key up.
4. Physics were being applied on input, which is bad for performance I believe. Updated this to store the state and apply it on `FixedUpdate()`.
5. I was not scaling physics with `Time.deltaTime`, doing so now should remove framerate dependence.

Beyond those fixes, I changed pitch and yaw to use the mouse. It is treated as an X and Y axes, with the `Inverted` processor used to apply the negative values.

I also add a gizmo to represent rotational momentum, useful for debugging flight controls.

Last minor change was moving the script to the StarFighter prefab rather than the individual ship. This allows options light the gizmo offset to be provided on a prefab basis, where the ship model is constant.