Refactors momentum gizmo into a separate `MonoBehaviour`.
This gizmo applies to any object with physics applied, so it didn't need to be limited to `StarFighters`. This also makes each class smaller and more self-contained.

The implementation is mostly the same, I just added a sphere to the gizmo, so it would display something small, even if the object was stationary.

I also used `[RequireComponent()]` to enforce that a `RigidBody` is provided.