Adds maximum mouse delta.
This limits mouse movement in a single frame to 50 pixels and emits the value as a linearly scaled integer within [-1. 1]. This brings the pitch and yaw values in `PilotAction` in line with the roll and thrust values.

This is somewhat misleading, because a typical piloting joystick moves in a circle, where the sum of the pitch and yaw components would be no more than 1 (of course a typical joystick would apply to the roll, not the yaw, but that's not important here). That joystick works as a circle, while this implementation is actually a square.

I think this is reasonable for now because generally have a faster pitch than a yaw, so it would be reasonable for these two be separate components. It also still does not allow for a user to "cheat" by turning faster diagonally, since the net result is a 1 value for both pitch and yaw.